Get things running again * Stop using LLVM due to http://hackage.haskell.org/trac/ghc/ticket/6158 (workaround fails) * Add strictness and new compile flags for marginal performance gains * Add convenience scripts

+++ Search.hs
@@ -9,7 +9,7 @@ import Types
 import Player
 import GameState
 
-minimax :: Int -> GameState -> [Float]
+minimax :: Int -> GameState -> [Double]
 minimax 0 node      = heuristicScores node
 minimax depth node  = case getChildren node of 
     []          -> finalScores node
